+++ tests/client_tests/like_post.py
@@ -1,0 +1,21 @@
+import os
+import json
+from jsonrpcclient import request
+from session_challenge import answer_challenge
+
+with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data', 'client_info.json'), 'r') as f:
+    client_info = json.load(f)
+with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data', 'session.social_session'), 'r') as f:
+    session_info = json.load(f)
+
+response = request(
+    client_info["endpoint"], "timeline.like_post",
+     session_identification={
+        "session_id": session_info["id"],
+        "client_public_hash": client_info["public_hash"],
+        "challenge_answer": answer_challenge(client_info["private_hash"], session_info["challenge"])
+     },
+     post_id="e3c158bfc9495b94f7f768662b1cea58-08e1344447946ccee8f60bb40384346b0b5d2bceffc28c9a6691c7f3f6f355e7"
+)
+
+print(response.data.result)