+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### v1.0.0.1 Alpha
  - Added method timeline.compose_post
+ - Added method timeline.like_post